# api-spec.yaml
openapi: 3.1.0
info:
  title: Authentik Role UI Backend
  version: 1.0.0
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    UserLink:
      type: object
      required: [username, name]
      properties:
        username:
          type: string
        name:
          type: string
    GroupLink:
      type: object
      required: [name]
      properties:
        name:
          type: string
    LoginAccount:
      type: object
      required: [kind, address]
      properties:
        kind:
          type: string
          description: e.g. "email", "telegram", "google"
        address:
          type: string
...
    UserGroupRoleSplit:
      description: Groups a user belongs to, split by their role in each group.
      type: object
      required: [leader, manager, member]
      properties:
        leader:
          type: array
          items:
            $ref: '#/components/schemas/GroupLink'
        manager:
          type: array
          items:
            $ref: '#/components/schemas/GroupLink'
        member:
          type: array
          items:
            $ref: '#/components/schemas/GroupLink'
    User:
      type: object
      required: [username, name, is_active, is_superuser, name_frozen, logins, groups, attributes]
      properties:
        username:
          type: string
        name:
          type: string
        is_active:
          type: boolean
        is_superuser:
          type: boolean
        name_frozen:
          type: boolean
          description: True after the user has used their one-time display-name change.
        logins:
          type: array
          items:
            $ref: '#/components/schemas/LoginAccount'
        groups:
          $ref: '#/components/schemas/UserGroupRoleSplit'
        attributes:
          type: array
          items:
            type: array
            prefixItems:
              - type: string
              - type: string
            minItems: 2
            maxItems: 2
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Caller lacks the required role in this group.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: User or group not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    BadRequest:
      description: Invalid request parameters or body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  parameters:
...
    Username:
      name: username
      in: path
      required: true
      schema:
        type: string
...
paths:
...
  /users:
    get:
      operationId: searchUsers
      summary: Search users
      description: >
        Returns users whose username, display name, or login addresses match the
        `search` regex. Returns an empty array when `search` is absent or blank.
      parameters:
        - name: search
          in: query
          required: false
          description: Regex pattern to match against username, name, and login addresses.
          schema:
            type: string
      responses:
        '200':
          description: Matching users (may be empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserLink'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me:
    get:
      operationId: getMe
      summary: Get caller's full profile
      responses:
        '200':
          description: Full profile of the authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/attributes:
    patch:
      operationId: patchMyAttributes
      summary: Replace own user-defined attributes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [attributes]
              properties:
                attributes:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        '200':
          description: Attributes updated.
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/name:
    put:
      operationId: setDisplayName
      summary: Change own display name (one-time)
      description: >
        Changes the caller's display name. Allowed once per user; blocked with
        403 if `name_frozen` is already set. Superusers can change their name
        any number of times regardless of the flag.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Display name updated.
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /users/{username}:
    get:
      operationId: getUser
      summary: Get a user by username
      parameters:
        - $ref: '#/components/parameters/Username'
      responses:
        '200':
          description: Full user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Groups
  # ---------------------------------------------------------------------------
...